test: fix pointer-to-int-cast warnings in TEST_CHECK_EQ
TEST_CHECK_EQ cast pointer arguments directly to long long, which
triggers -Wpointer-to-int-cast on 32-bit builds since long long is a
different size than a pointer. Route through intptr_t (matching
pointer width) before widening to long long to silence the warning
without truncating values.

TEST_CHECK_U64_EQ is left unchanged since it's used with genuine
64-bit integers (file sizes, FILETIMEs), not pointers.

+++ test/test_assert.h
@@ -1,6 +1,7 @@
 #ifndef WIBO_TEST_ASSERT_H
 #define WIBO_TEST_ASSERT_H
 
+#include <stdint.h>
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
@@ -27,8 +28,8 @@
 
 #define TEST_CHECK_EQ(expected, actual) \
     do { \
-        long long _expected_value = (long long)(expected); \
-        long long _actual_value = (long long)(actual); \
+        long long _expected_value = (long long)(intptr_t)(expected); \
+        long long _actual_value = (long long)(intptr_t)(actual); \
         if (_expected_value != _actual_value) { \
             TEST_FAIL("Expected %s (%lld) == %s (%lld)", \
                       #expected, _expected_value, #actual, _actual_value); \
